backend/requirements.txt
flask
flask-cors
pymongo
bcrypt
opencv-python
deepface
numpy
python-dotenv